requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
crc16==0.1.1
Django==4.1.3
django-bitfield==2.2.0
django-cors-headers==3.13.0
django-filter==22.1
django-geo==0.8
django-geojson==4.0.0
django-js-asset==2.0.0
django-leaflet==0.28.3
django-mptt==0.14.0
django-rest-swagger==2.2.0
django-separatedvaluesfield==0.6.0
django-timezone-field==5.0
django-uuidfield==0.5.0
djangorestframework==3.14.0
djangorestframework-gis @ https://github.com/openwisp/django-rest-framework-gis/tarball/master
Faker==12.0.1
GDAL @ file:///C:/Users/Clevo/Downloads/django-leaflet-master/GDAL-3.4.3-cp38-cp38-win_amd64.whl
idna==3.4
importlib-metadata==5.1.0
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
mixer==7.2.2
nmea-parser==0.5.0
openapi-codec==1.3.2
Pillow==9.3.0
psycopg2-binary==2.9.5
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
simplejson==3.18.0
six==1.16.0
South==1.0.2
sqlparse==0.4.3
tzdata==2022.6
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.11.0
